fix test requirements install
