+++ sdk/python/requirements-ci.txt
@@ -4,11 +4,6 @@ black==19.10b0
 isort>=5
 grpcio-tools
 mypy-protobuf
-pytest
-pytest-lazy-fixture==0.6.3
-pytest-mock
-pytest-timeout
-pytest-ordering==0.6.*
 pyspark==2.4.2
 pandas~=1.0.0
 mock==2.0.0
