+++ tests/requirements.txt
@@ -1,5 +1,8 @@
-mock==2.0.0
 pytest==6.0.0
+pytest-lazy-fixture==0.6.3
+pytest-mock
+pytest-timeout
+pytest-ordering==0.6.*
 pytest-benchmark==3.2.2
 pytest-mock==1.10.4
 pytest-timeout==1.3.3
